+++ apps/web/src/components/SearchInput.tsx
